Improved Agreements Template
Instead of displaying an "ugly" link [Show] beneath the label of a
checkbox, it is possible to emphasize part of the label text with square
braces. If those are found, the part of the text between the braces will
be replaced through the link and the link [Show] disappears.

Example:
I agree to the [terms and conditions].

Checkbox label:
I agree to the terms and conditions.
Where "terms and conditions" is a link, opening the content in a new
window. The link also works with the middle button of the mouse or if
context menu for opening the link in a new tab is used.

More info on this improvement:
http://magento.xonu.de/tutorials/button-loesung-optimierung/